+++ resources/js/components/gallery/photoModule/PhotoRatingWidget.vue
@@ -23,10 +23,10 @@
 				:disabled="loading || props.rating.rating_user === 0"
 				:class="{
 					'px-2 py-1 text-sm rounded transition-colors': true,
-					'bg-red-600 hover:bg-red-700 text-white': props.rating.rating_user > 0 && !loading,
+					'bg-red-600 hover:bg-red-700 text-white hover:cursor-pointer': props.rating.rating_user > 0 && !loading,
 					'bg-gray-500 text-gray-300 cursor-not-allowed': loading || props.rating.rating_user === 0,
 				}"
-				@click="handleRatingClick(0)"
+				@click="handleRatingClick(props.photoId, 0)"
 			>
 				×
 			</button>
@@ -43,7 +43,7 @@
 				}"
 				@mouseenter="handleMouseEnter(rating)"
 				@mouseleave="handleMouseLeave()"
-				@click="handleRatingClick(rating as 1 | 2 | 3 | 4 | 5)"
+				@click="handleRatingClick(props.photoId, rating as 1 | 2 | 3 | 4 | 5)"
 			>
 				<i
 					:class="{
@@ -63,21 +63,19 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
 import { useLycheeStateStore } from "@/stores/LycheeState";
 import { usePhotoStore } from "@/stores/PhotoState";
 import { useToast } from "primevue/usetoast";
-import PhotoService from "@/services/photo-service";
 import { useUserStore } from "@/stores/UserState";
 import StarRow from "@/components/icons/StarRow.vue";
+import { useRating } from "@/composables/photo/useRating";
 
 const lycheeStore = useLycheeStateStore();
 const photoStore = usePhotoStore();
 const userStore = useUserStore();
 const toast = useToast();
 
-const loading = ref(false);
-const hoverRating = ref<number | null>(null);
+const { hoverRating, loading, handleRatingClick } = useRating(photoStore, toast, userStore);
 
 const props = defineProps<{
 	photoId: string;
@@ -92,53 +90,5 @@ function handleMouseEnter(rating: number) {
 
 function handleMouseLeave() {
 	hoverRating.value = null;
-}
-
-function handleRatingClick(rating: 0 | 1 | 2 | 3 | 4 | 5) {
-	if (loading.value) {
-		return;
-	}
-
-	loading.value = true;
-
-	PhotoService.setRating(props.photoId, rating)
-		.then((response) => {
-			// Update photo store with new photo data (includes updated rating)
-			photoStore.photo = response.data;
-
-			// Show success message
-			const message = rating === 0 ? "gallery.photo.rating.removed" : "gallery.photo.rating.saved";
-			toast.add({
-				severity: "success",
-				summary: "Success",
-				detail: message,
-				life: 3000,
-			});
-		})
-		.catch((error) => {
-			console.error("Failed to save rating:", error);
-
-			// Show error toast
-			let errorMessage = "gallery.photo.rating.error";
-
-			if (error.response?.status === 401) {
-				errorMessage = "gallery.photo.rating.error_unauthorized";
-			} else if (error.response?.status === 403) {
-				errorMessage = "gallery.photo.rating.error_forbidden";
-			} else if (error.response?.status === 404) {
-				errorMessage = "gallery.photo.rating.error_not_found";
-			}
-
-			toast.add({
-				severity: "error",
-				summary: "Error",
-				detail: errorMessage,
-				life: 5000,
-			});
-		})
-		.finally(() => {
-			loading.value = false;
-			hoverRating.value = null;
-		});
 }
 </script>
